fix model Post

+++ shareJourney/journeys/models.py
@@ -52,8 +52,8 @@ class Participation(Interaction):  # ds user tham gia hành trình
 class Post(Interaction):
     content = models.TextField()
     visit_point = models.CharField(max_length=255, blank=True, null=True)
-    latitude = models.DecimalField(max_digits=9, decimal_places=6, blank=True, null=True)
-    longitude = models.DecimalField(max_digits=9, decimal_places=6, blank=True, null=True)
+    latitude = models.FloatField(blank=True, null=True)
+    longitude = models.FloatField(blank=True, null=True)
     estimated_time_of_arrival = models.CharField(max_length=100, blank=True, null=True)
